Add dh_strip_nondeterminism
This is a self-contained Perl script that should hopefully be suitable
for inclusion in debhelper itself.  I'm putting it in this repository
temporarily but ultimately it should be added to the reproducible
debhelper repo and a patch should be submitted to the BTS for debhelper.

For now supports gzip only.